ok done
spoon cursor and new approach to the leaf reading. decided to add another layer to the tea leaf reading: the tea leaf text wasn't working the way i wanted to so instead, im making them form various patterns in the cup and then creating an array of interpretations of the patterns, along with an API quote.